Moved `LayersAdapter` class to separate file
